+++ src/Infrastructure/Translation/TranslatorInterface.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+
+namespace Iriven\PhpFormGenerator\Infrastructure\Translation;
+
+interface TranslatorInterface
+{
+    public function trans(string $key, array $parameters = []): string;
+}